requirements.txt
streamlit
requests
sentence-transformers
scikit-learn
torch
pandas
fastapi
uvicorn
reportlab
faiss-cpu